requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
backoff==2.2.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.52
cryptography==42.0.5
fastavro==1.9.4
Flask==3.0.2
Flask-WTF==1.2.1
frozenlist==1.4.1
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
pdfminer==20191125
pdfminer.six==20231228
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
python-dotenv==1.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
